source/lambda/requirements.txt
boto3
google-cloud-pubsub
google-api-core
google-auth
